rename empty_node to nonexistence_proof
